restore user

+++ content/modul.php
@@ -8,12 +8,14 @@ $id_role = isset($_SESSION['ID_ROLE']) ? $_SESSION['ID_ROLE'] : '';
 // ON untuk memberitahu secara spesifik data mana yang akan dihubungkan
 
 $rowStudent = mysqli_fetch_assoc(mysqli_query($config, "SELECT * FROM students WHERE id = '$id_user'"));
-$id_major = $rowStudent['id_major'];
+$id_major = isset($rowStudent['id_major']) ? $rowStudent['id_major'] : '';
 
-if($id_role == 2){
+if($id_role == 8){
     $where = "WHERE moduls.id_major='$id_major'";
-}elseif($id_role == 1){
+}elseif($id_role == 9){
     $where = "WHERE moduls.id_instructor='$id_user'";
+}else{
+    $where = "";
 }
 
 $queryModul = mysqli_query($config, "SELECT majors.name AS major_name, instructors.name AS instructor_name, moduls. * FROM moduls LEFT JOIN majors ON majors.id = moduls.id_major LEFT JOIN instructors ON instructors.id = moduls.id_instructor $where ORDER BY moduls.id DESC");
@@ -25,7 +27,7 @@ $rowModul = mysqli_fetch_all($queryModul, MYSQLI_ASSOC);
         <div class="card">
             <div class="card-body">
                 <h5 class="card-title">Data Modul</h5>
-                <?php if($_SESSION['ID_ROLE'] == 1): ?>
+                <?php if(canAddModul($id_role)): ?>
                     <div class="mb-3" align="right">
                         <a href="?page=tambah-modul" class="btn btn-primary">Add Modul</a>
                     </div>
@@ -57,9 +59,11 @@ $rowModul = mysqli_fetch_all($queryModul, MYSQLI_ASSOC);
                                         <td><?= $data['instructor_name'] ?></td>
                                         <td><?= $data['major_name'] ?></td>
                                         <td>
-                                            <!-- <a href="?page=tambah-modul&edit=<?php echo $data['id'] ?>" class="btn btn-warning btn-sm">Edit</a> -->
-                                            <a onclick="return confirm('Are you sure??')"
-                                                href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-danger btn-sm">Delete</a>
+                                            <?php if ($id_role == 1): ?>
+                                                <!-- <a href="?page=tambah-modul&edit=<?php echo $data['id'] ?>" class="btn btn-warning btn-sm">Edit</a> -->
+                                                <a onclick="return confirm('Are you sure??')"
+                                                    href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-danger btn-sm">Delete</a>
+                                            <?php endif ?>
                                         </td>
                                     </tr>
                             <?php endforeach ?>
